licensing: move accesstoken back to licensing
