--- internal/accesstoken/BUILD.bazel
@@ -1,24 +1,0 @@
-load("//dev:go_defs.bzl", "go_test")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "accesstoken",
-    srcs = ["access_token.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/accesstoken",
-    visibility = ["//:__subpackages__"],
-    deps = [
-        "//internal/hashutil",
-        "//lib/errors",
-    ],
-)
-
-go_test(
-    name = "accesstoken_test",
-    srcs = ["access_token_test.go"],
-    embed = [":accesstoken"],
-    deps = [
-        "@com_github_hexops_autogold_v2//:autogold",
-        "@com_github_stretchr_testify//assert",
-        "@com_github_stretchr_testify//require",
-    ],
-)
