+++ internal/licensing/BUILD.bazel
@@ -4,6 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "licensing",
     srcs = [
+        "access_token.go",
         "check.go",
         "codygateway.go",
         "conf.go",
@@ -19,12 +20,12 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/licensing",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/accesstoken",
         "//internal/conf",
         "//internal/conf/conftypes",
         "//internal/env",
         "//internal/errcode",
         "//internal/goroutine",
+        "//internal/hashutil",
         "//internal/httpcli",
         "//internal/license",
         "//internal/redispool",
@@ -41,6 +42,7 @@ go_test(
     name = "licensing_test",
     timeout = "short",
     srcs = [
+        "access_token_test.go",
         "check_test.go",
         "codygateway_test.go",
         "features_test.go",
@@ -59,6 +61,7 @@ go_test(
         "@com_github_derision_test_glock//:glock",
         "@com_github_gomodule_redigo//redis",
         "@com_github_google_go_cmp//cmp",
+        "@com_github_hexops_autogold_v2//:autogold",
         "@com_github_sourcegraph_log//logtest",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
