+++ internal/licensing/check.go
@@ -2,18 +2,15 @@ package licensing
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
-	"time"
-
 	"net/http"
 	"net/url"
+	"time"
 
 	"github.com/derision-test/glock"
 	"github.com/sourcegraph/log"
 
-	"context"
-
-	licensing "github.com/sourcegraph/sourcegraph/internal/accesstoken"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
@@ -176,7 +173,7 @@ func StartLicenseCheck(originalCtx context.Context, logger log.Logger, siteID st
 		ctxWithCancel, cancel = context.WithCancel(originalCtx)
 
 		prevLicenseToken, _ := store.Get(prevLicenseTokenKey).String()
-		licenseToken := licensing.GenerateLicenseKeyBasedAccessToken(conf.Get().LicenseKey)
+		licenseToken := GenerateLicenseKeyBasedAccessToken(conf.Get().LicenseKey)
 		var initialWaitInterval time.Duration = 0
 		if prevLicenseToken == licenseToken {
 			initialWaitInterval, _ = calcDurationSinceLastCalled(glock.NewRealClock())
